Implemented the mapper classes of Comment, Post, Subreddit and Vote

+++ Reddit-backend/src/main/java/com/example/reddit/comment/CommentDto.java
@@ -14,7 +14,7 @@ public record CommentDto(
         String text,
         Integer score,
         Set<VoteDto> votes,
-        Set<Integer> savedBy,
+        Set<Long> savedBy,
         Set<CommentDto> replies,
         Long parentId,
         Long postId) {
